Move cheat packets to their own functions.
